+++ sources/packages/web/src/components/students/RestrictionBanner.vue
