+++ sources/packages/web/src/store/modules/student/actions.ts
@@ -1,6 +1,6 @@
 import { ActionContext, ActionTree } from "vuex";
 import { RootState } from "@/types";
-import { StudentState } from "./student";
+import { StudentState, StudentRestriction } from "./student";
 import { StudentService } from "@/services/StudentService";
 
 export const actions: ActionTree<StudentState, RootState> = {
@@ -21,7 +21,13 @@ export const actions: ActionTree<StudentState, RootState> = {
   async updateStudentRestrictions(
     context: ActionContext<StudentState, RootState>,
   ) {
-    const restrictions = await StudentService.shared.getStudentRestriction();
+    const response = await StudentService.shared.getStudentRestriction();
+    const restrictions = response
+      ? response.map<StudentRestriction>((restriction) => ({
+          code: restriction.code,
+          type: restriction.type,
+        }))
+      : [];
     context.commit("updateStudentRestrictions", restrictions);
   },
 };
