+++ sources/packages/web/src/store/modules/student/getters.ts
@@ -1,10 +1,6 @@
 import { GetterTree } from "vuex";
-import {
-  RootState,
-  RestrictionNotificationType,
-  StudentRestriction,
-} from "@/types";
-import { StudentState } from "./student";
+import { RootState, RestrictionNotificationType } from "@/types";
+import { StudentState, StudentRestriction } from "./student";
 
 export const getters: GetterTree<StudentState, RootState> = {
   hasRestrictionError(state: StudentState): boolean {
@@ -17,7 +13,12 @@ export const getters: GetterTree<StudentState, RootState> = {
       (restriction) => restriction.type === RestrictionNotificationType.Warning,
     );
   },
+  //We are considering the restrictions of notification type warning or error only.
   getRestrictions(state: StudentState): StudentRestriction[] {
-    return state.restrictions;
+    return state.restrictions.filter(
+      (restriction) =>
+        restriction.type === RestrictionNotificationType.Warning ||
+        restriction.type === RestrictionNotificationType.Error,
+    );
   },
 };
