--- sources/packages/web/src/views/student/StudentAccountActivityFormModel.ts
@@ -1,4 +1,0 @@
-import { StudentRestriction } from "@/types";
-export interface StudentAccountActivityFormModel {
-  restrictions: StudentRestriction[];
-}
